Issue #108 LambdaUtils dependencies
- Single source file was rather large - broke up into multiple.
