+++ lambda-utils/tsconfig.json
@@ -21,5 +21,6 @@
     "lib": [
       "es2017"
     ]
-  }
+  },
+  "exclude" : ["lib/*.test.ts"]
 }
